requirements.txt
appdirs==1.4.4
asttokens==2.4.1
black==24.2.0
click==8.1.7
contourpy==1.2.0
cupy-cuda12x==13.0.0
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastrlock==0.8.2
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
ipython==8.22.1
jedi==0.19.1
Jinja2==3.1.3
kiwisolver==1.4.5
llvmlite==0.42.0
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.2.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycuda==2024.1
Pygments==2.17.2
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
pytools==2023.1.1
pytz==2024.1
ruff==0.3.3
scipy==1.12.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
tomli==2.0.1
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
traitlets==5.14.1
triton==2.2.0
typing_extensions==4.9.0
tzdata==2024.1
wcwidth==0.2.13
